- WorldServer/ChannelServer were desynced in terms of player info update packets, I've made them consistent  - Added explicit types to many different enumerations  - Fixed PacketReader which had a getHeader implementation that only worked as long as header_t was int16_t  - Changed header_t to uint16_t  - Moved ServerTypes into VanaConstants.h
git-svn-id: https://dev.chiasoft.net:8443/svn/svn_vana/branches/buis_vana_overhaul@3110 23f3cdc0-e6d1-42be-a0c1-51f4125517bb

+++ src/ChannelServer/PlayerDataProvider.cpp
@@ -152,7 +152,7 @@ void PlayerDataProvider::sendPacket(PacketCreator &packet, int32_t minGmLevel) {
 void PlayerDataProvider::updatePlayer(PacketReader &packet) {
 	int32_t playerId = packet.get<int32_t>();
 	if (PlayerData *player = getPlayerData(playerId)) {
-		int32_t updateBits = packet.get<int32_t>();
+		int8_t updateBits = packet.get<int8_t>();
 		bool updateParty = false;
 		bool updateGuild = false;
 		bool updateAlliance = false;
